requirements.txt
datasets
requests
openai
python-dotenv~=1.0.1
pandas
semantic-kernel>=1.29.0
semantic-kernel[azure]
semantic-kernel[mcp]